Move TwoFingerTracker to separate file and reuse metersPerUnit in toCentimeters
